Move network stuff to new folder "network"
