+++ src/network/network.cpp
@@ -1,5 +1,5 @@
 #include "network.h"
-#include "util/util.h"
+#include "../util/util.h"
 
 #include <errno.h>
 #include <cstring>
